\name{DBFEr-package}
\alias{DBFEr-package}
\alias{DBFEr}
\docType{package}
\title{
Distribution Based Feature Extraction in R
}
\description{
This package extract features based on distribution of lengths.
}

\references{
Maciej P. et al. DBFE: distribution-based feature extraction from structural variants in whole-genome data, Bioinformatics, Volume 38, Issue 19, 1 October, Pages 4466–4473(2022). https://doi.org/10.1093/bioinformatics/btac513
}
\author{
BIOSTAT615 Group8
}
