% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionBasedFeatureExtractor.R
\name{DistributionBasedFeatureExtractor}
\alias{DistributionBasedFeatureExtractor}
\title{Extract Features from Distribution of Lengths}
\usage{
DistributionBasedFeatureExtractor(
  values,
  labels,
  breakpoint_type = "quantile",
  n_bins = 4,
  prefix = "dbfe",
  bw = 0.5,
  resolution = 200,
  log_scale = TRUE,
  value_range = NULL
)
}
\arguments{
\item{values}{A dataframe of lengths to be discretized with 2 columns, sample_id and lengths.}

\item{labels}{A dataframe of labels with 2 columns, sample_id and labels. Only needed for supervised scenarios.}

\item{breakpoint_type}{A string of type of breakpoints to generate.
Available options are: equal, quantile, supervised, clustering. Defaults to 'quantile'.}

\item{n_bins}{An integer or "auto" of number of features to generate.
An integer explicitly selects given number of features. "auto" is only for supervised.}

\item{prefix}{A string of the name}

\item{bw}{A float number of bandwidth for kde. Used only in supervised breakpoints. Defaults to 0.5.}

\item{resolution}{An integer of resolution of kde. Used only in supervised breakpoints. Defaults to 200.}

\item{log_scale}{A bool whether log scale should be used on the x axis to calculate the bins.
Used only in supervised breakpoints. Defaults to True.}

\item{value_range}{Range of values. NULL is allowed.}
}
\value{
A dataframe with all the features, combined with labels for supervised scenarios.
}
\description{
4 feature extraction methods that facilitate classifier learning
from genomic variants are included in this function.
}
\examples{
\dontrun{
DistributionBasedFeatureExtractor(values=df, labels=y)
}
}
